webserver: Refactor validation to allow reuse
Refactor the validation of GET /chat/ to allow reuse by new endpoints:

 - A new validate.ts has a few routines for setting up exception
   catching validation and handlers which the chatRouter now uses,
   reducing boilerplate.

 - GET /chat/ validation is split into the body & player validation
   (checkPlayerIntegrity in validate.ts) and the GET /chat/ specific
   "message" field (checkChatIntegrity in postChat.ts).
